// src/lib/api/generated/models/bankaBankV1Currency.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum bankaBankV1Currency {
    CURRENCY_UNSPECIFIED = 'CURRENCY_UNSPECIFIED',
    CURRENCY_RSD = 'CURRENCY_RSD',
    CURRENCY_EUR = 'CURRENCY_EUR',
    CURRENCY_CHF = 'CURRENCY_CHF',
    CURRENCY_USD = 'CURRENCY_USD',
    CURRENCY_GBP = 'CURRENCY_GBP',
    CURRENCY_JPY = 'CURRENCY_JPY',
    CURRENCY_CAD = 'CURRENCY_CAD',
    CURRENCY_AUD = 'CURRENCY_AUD',
}