// src/lib/api/generated/models/v1CreateForexForwardRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { bankaBankV1Currency } from './bankaBankV1Currency';
export type v1CreateForexForwardRequest = {
    baseCurrency?: bankaBankV1Currency;
    notional?: string;
    settlementDate?: string;
};